% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest}
\alias{bucharest}
\title{rcoins test data}
\format{
An object of class \code{list} of length 3.
}
\source{
OpenStreetMap
}
\usage{
bucharest
}
\description{
Data extracted from OpenStreetMap for testing the rcoins package.
}
\keyword{datasets}
