% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Get example data}
\usage{
get_example_data()
}
\value{
A list of sf objects containing the OSM data.
}
\description{
This function retrieves example OpenStreetMap (OSM) data for the city of
Bucharest, Romania, from a persistent URL on the 4TU.ResearchData data
repository. The dataset includes the street network and the geometry of the
Dâmbovița river.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_example_data()
\dontshow{\}) # examplesIf}
}
