% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke.R
\name{stroke}
\alias{stroke}
\title{Identify continuous lines in a network}
\usage{
stroke(
  edges,
  angle_threshold = 0,
  attributes = FALSE,
  flow_mode = FALSE,
  from_edge = NULL
)
}
\arguments{
\item{edges}{An object of class \code{\link[sfc]{sfc}} (or compatible),
including the edge geometries (should be of type LineString or
MultiLineString).}

\item{angle_threshold}{Consecutive line segments can be considered part of
the same stroke if the internal angle they form is larger than
\code{angle_threshold} (in degrees). It should fall in the range
\eqn{0 \leq angle_threshold < 180}.}

\item{attributes}{If \code{TRUE}, return a label for each edge, representing
the groups each edge belongs to. Only possible for \code{flow_mode = TRUE}.}

\item{flow_mode}{If \code{TRUE}, line segments that belong to the same edge
are not split across strokes (even if they form internal angles smaller than
\code{angle_threshold}).}

\item{from_edge}{Only look for the continuous strokes that include the
provided edges or line segments.}
}
\value{
An object of class \code{\link[sfc]{sfc}} (if
\code{attributes = FALSE}), a vector with the same length as \code{edges}
otherwise.
}
\description{
Apply the Continuity in Street Network (COINS) method to identify
sequences of edges that form naturally continuous strokes in a network.
}
